Added value checker
In a deck of playing cards, there are 4 times a value that can be picked.
The game caps this so if a number is generated so now there are 5 of the same value, the value will be randomised so a new value will be picked. Two conditions are used to check and must both be valid to use the value.
As tens, jacks, queens and kings all count as 10, there is a chance the card can be picked 12 times. Similar logic is used from the condition above to implement this condition. 

To check this, the function valueCount is used to filter and then count the number of times the value appears. This is returned for use. 

If the card meets the first or second if statements it calls the function getNewCard which then recalls newCard in order to get a new card.